+++ packages/payment-integration-api/src/form/form-fields.ts
@@ -32,6 +32,7 @@ export default interface FormField {
     secret?: boolean;
     min?: string | number;
     max?: string | number;
+    inputDateFormat?: string;
     options?: FormFieldOptions;
     requirements?: CustomerPasswordRequirements;
 }