+++ packages/payment-integration-api/src/index.ts
@@ -110,8 +110,10 @@ export {
     SubmitOrderAction,
 } from './order';
 export {
-    BlueSnapDirectEcpInstrument,
+    EcpInstrument,
     BlueSnapDirectEcpPayload,
+    SepaInstrument,
+    BlueSnapDirectSepaPayload,
     CardInstrument,
     CreditCardInstrument,
     WithBankAccountInstrument,
@@ -136,6 +138,9 @@ export {
     VaultedInstrument,
     WithAccountCreation,
     StripeV3FormattedPayload,
+    PaypalInstrument,
+    FormattedPayload,
+    isHostedVaultedInstrument,
 } from './payment';
 export { default as PaymentIntegrationSelectors } from './payment-integration-selectors';
 export { default as PaymentIntegrationService } from './payment-integration-service';