+++ packages/paypal-commerce-integration/src/paypal-commerce-credit/paypal-commerce-credit-customer-strategy.spec.ts
@@ -43,7 +43,8 @@ describe('PayPalCommerceCreditCustomerStrategy', () => {
 
     const methodId = 'paypalcommercecredit';
     const defaultContainerId = 'paypal-commerce-credit-container-mock-id';
-    const approveDataOrderId = 'ORDER_ID';
+    const paypalOrderId = 'ORDER_ID';
+    const approveDataOrderId = paypalOrderId;
 
     const paypalCommerceCreditOptions: PayPalCommerceCreditCustomerInitializeOptions = {
         container: defaultContainerId,
@@ -55,6 +56,24 @@ describe('PayPalCommerceCreditCustomerStrategy', () => {
         WithPayPalCommerceCreditCustomerInitializeOptions = {
         methodId,
         paypalcommercecredit: paypalCommerceCreditOptions,
+    };
+
+    const paypalShippingAddressPayloadMock = {
+        city: 'New York',
+        country_code: 'US',
+        postal_code: '07564',
+        state: 'New York',
+    };
+
+    const paypalSelectedShippingOptionPayloadMock = {
+        amount: {
+            currency_code: 'USD',
+            value: '100',
+        },
+        id: '1',
+        label: 'Free shipping',
+        selected: true,
+        type: 'type_shipping',
     };
 
     beforeEach(() => {
@@ -135,34 +154,12 @@ describe('PayPalCommerceCreditCustomerStrategy', () => {
                     }
                 });
 
-                eventEmitter.on('onShippingAddressChange', () => {
-                    if (options.onShippingAddressChange) {
-                        options.onShippingAddressChange({
-                            orderId: approveDataOrderId,
-                            shippingAddress: {
-                                city: 'New York',
-                                country_code: 'US',
-                                postal_code: '07564',
-                                state: 'New York',
-                            },
-                        });
-                    }
-                });
-
-                eventEmitter.on('onShippingOptionsChange', () => {
-                    if (options.onShippingOptionsChange) {
-                        options.onShippingOptionsChange({
-                            orderId: approveDataOrderId,
-                            selectedShippingOption: {
-                                amount: {
-                                    currency_code: 'USD',
-                                    value: '100',
-                                },
-                                id: '1',
-                                label: 'Free shipping',
-                                selected: true,
-                                type: 'type_shipping',
-                            },
+                eventEmitter.on('onShippingChange', () => {
+                    if (options.onShippingChange) {
+                        options.onShippingChange({
+                            orderID: paypalOrderId,
+                            shipping_address: paypalShippingAddressPayloadMock,
+                            selected_shipping_option: paypalSelectedShippingOptionPayloadMock,
                         });
                     }
                 });
@@ -284,8 +281,7 @@ describe('PayPalCommerceCreditCustomerStrategy', () => {
                     label: 'checkout',
                 },
                 createOrder: expect.any(Function),
-                onShippingAddressChange: expect.any(Function),
-                onShippingOptionsChange: expect.any(Function),
+                onShippingChange: expect.any(Function),
                 onApprove: expect.any(Function),
                 onClick: expect.any(Function),
             });
@@ -517,7 +513,7 @@ describe('PayPalCommerceCreditCustomerStrategy', () => {
         });
     });
 
-    describe('#onShippingAddressChange button callback', () => {
+    describe('#onShippingChange button callback', () => {
         beforeEach(() => {
             const paymentMethodWithShippingOptionsFeature = {
                 ...paymentMethod,
@@ -554,7 +550,7 @@ describe('PayPalCommerceCreditCustomerStrategy', () => {
 
             await strategy.initialize(initializationOptions);
 
-            eventEmitter.emit('onShippingAddressChange');
+            eventEmitter.emit('onShippingChange');
 
             await new Promise((resolve) => process.nextTick(resolve));
 
@@ -565,7 +561,7 @@ describe('PayPalCommerceCreditCustomerStrategy', () => {
         it('selects shipping option after address update', async () => {
             await strategy.initialize(initializationOptions);
 
-            eventEmitter.emit('onShippingAddressChange');
+            eventEmitter.emit('onShippingChange');
 
             await new Promise((resolve) => process.nextTick(resolve));
 
@@ -586,47 +582,7 @@ describe('PayPalCommerceCreditCustomerStrategy', () => {
 
             await strategy.initialize(initializationOptions);
 
-            eventEmitter.emit('onShippingAddressChange');
-
-            await new Promise((resolve) => process.nextTick(resolve));
-
-            expect(paypalCommerceIntegrationService.updateOrder).toHaveBeenCalled();
-        });
-    });
-
-    describe('#onShippingOptionsChange button callback', () => {
-        beforeEach(() => {
-            const paymentMethodWithShippingOptionsFeature = {
-                ...paymentMethod,
-                initializationData: {
-                    ...paymentMethod.initializationData,
-                    isHostedCheckoutEnabled: true,
-                },
-            };
-
-            jest.spyOn(
-                paymentIntegrationService.getState(),
-                'getPaymentMethodOrThrow',
-            ).mockReturnValue(paymentMethodWithShippingOptionsFeature);
-        });
-
-        it('selects shipping option', async () => {
-            await strategy.initialize(initializationOptions);
-
-            eventEmitter.emit('onShippingOptionsChange');
-
-            await new Promise((resolve) => process.nextTick(resolve));
-
-            expect(paypalCommerceIntegrationService.getShippingOptionOrThrow).toHaveBeenCalled();
-            expect(paymentIntegrationService.selectShippingOption).toHaveBeenCalledWith(
-                getShippingOption().id,
-            );
-        });
-
-        it('updates PayPal order', async () => {
-            await strategy.initialize(initializationOptions);
-
-            eventEmitter.emit('onShippingOptionsChange');
+            eventEmitter.emit('onShippingChange');
 
             await new Promise((resolve) => process.nextTick(resolve));