+++ packages/paypal-commerce-integration/src/paypal-commerce/paypal-commerce-button-strategy.spec.ts
@@ -215,20 +215,12 @@ describe('PayPalCommerceButtonStrategy', () => {
                     }
                 });
 
-                eventEmitter.on('onShippingAddressChange', () => {
-                    if (options.onShippingAddressChange) {
-                        options.onShippingAddressChange({
-                            orderId: paypalOrderId,
-                            shippingAddress: paypalShippingAddressPayloadMock,
-                        });
-                    }
-                });
-
-                eventEmitter.on('onShippingOptionsChange', () => {
-                    if (options.onShippingOptionsChange) {
-                        options.onShippingOptionsChange({
-                            orderId: paypalOrderId,
-                            selectedShippingOption: paypalSelectedShippingOptionPayloadMock,
+                eventEmitter.on('onShippingChange', () => {
+                    if (options.onShippingChange) {
+                        options.onShippingChange({
+                            orderID: paypalOrderId,
+                            shipping_address: paypalShippingAddressPayloadMock,
+                            selected_shipping_option: paypalSelectedShippingOptionPayloadMock,
                         });
                     }
                 });
@@ -406,8 +398,7 @@ describe('PayPalCommerceButtonStrategy', () => {
                 fundingSource: paypalSdk.FUNDING.PAYPAL,
                 style: paypalCommerceOptions.style,
                 createOrder: expect.any(Function),
-                onShippingAddressChange: expect.any(Function),
-                onShippingOptionsChange: expect.any(Function),
+                onShippingChange: expect.any(Function),
                 onApprove: expect.any(Function),
             });
         });
@@ -644,7 +635,7 @@ describe('PayPalCommerceButtonStrategy', () => {
         });
     });
 
-    describe('#onShippingAddressChange button callback', () => {
+    describe('#onShippingChange button callback', () => {
         beforeEach(() => {
             const paymentMethodWithShippingOptionsFeature = {
                 ...paymentMethod,
@@ -681,7 +672,7 @@ describe('PayPalCommerceButtonStrategy', () => {
 
             await strategy.initialize(initializationOptions);
 
-            eventEmitter.emit('onShippingAddressChange');
+            eventEmitter.emit('onShippingChange');
 
             await new Promise((resolve) => process.nextTick(resolve));
 
@@ -692,7 +683,7 @@ describe('PayPalCommerceButtonStrategy', () => {
         it('selects shipping option after address update', async () => {
             await strategy.initialize(initializationOptions);
 
-            eventEmitter.emit('onShippingAddressChange');
+            eventEmitter.emit('onShippingChange');
 
             await new Promise((resolve) => process.nextTick(resolve));
 
@@ -713,47 +704,7 @@ describe('PayPalCommerceButtonStrategy', () => {
 
             await strategy.initialize(initializationOptions);
 
-            eventEmitter.emit('onShippingAddressChange');
-
-            await new Promise((resolve) => process.nextTick(resolve));
-
-            expect(paypalCommerceIntegrationService.updateOrder).toHaveBeenCalled();
-        });
-    });
-
-    describe('#onShippingOptionsChange button callback', () => {
-        beforeEach(() => {
-            const paymentMethodWithShippingOptionsFeature = {
-                ...paymentMethod,
-                initializationData: {
-                    ...paymentMethod.initializationData,
-                    isHostedCheckoutEnabled: true,
-                },
-            };
-
-            jest.spyOn(
-                paymentIntegrationService.getState(),
-                'getPaymentMethodOrThrow',
-            ).mockReturnValue(paymentMethodWithShippingOptionsFeature);
-        });
-
-        it('selects shipping option', async () => {
-            await strategy.initialize(initializationOptions);
-
-            eventEmitter.emit('onShippingOptionsChange');
-
-            await new Promise((resolve) => process.nextTick(resolve));
-
-            expect(paypalCommerceIntegrationService.getShippingOptionOrThrow).toHaveBeenCalled();
-            expect(paymentIntegrationService.selectShippingOption).toHaveBeenCalledWith(
-                getShippingOption().id,
-            );
-        });
-
-        it('updates PayPal order', async () => {
-            await strategy.initialize(initializationOptions);
-
-            eventEmitter.emit('onShippingOptionsChange');
+            eventEmitter.emit('onShippingChange');
 
             await new Promise((resolve) => process.nextTick(resolve));